covered fundamentals up to arrays

+++ 2.Var_constants_types/02_Calc/main.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 
 int main(int argc, char **argv) {
-	double subtotal;
+	int subtotal;
 	double tax;
 	double grandtotal;
 	double taxrate;
@@ -11,6 +11,6 @@ int main(int argc, char **argv) {
 	tax = subtotal * taxrate;
 	grandtotal = subtotal + tax;
 
-	printf( "The tax on %.2f is %.2f, so the grand total is %.2f.\n", subtotal, tax, grandtotal );
+	printf( "The tax on %d is %.2f, so the grand total is %.2f.\n", subtotal, tax, grandtotal );
 	return 0;
 }
